<style> :스타일 단순 수정
